requirements.txt
eth-brownie==1.17.2
numpy==1.22.1
prometheus-client==0.13.0
